major refactor of reduce into traits
