+++ crates/burn-wgpu/src/compute/tune_key.rs
@@ -11,10 +11,8 @@ use crate::fusion::FusionElemWiseAutotuneKey;
 pub enum JitAutotuneKey {
     /// Key for matmul operation
     Matmul(MatmulAutotuneKey),
-    /// Key for sum_dim operations
-    SumDim(ReduceAutotuneKey),
-    /// Key for mean_dim operations
-    MeanDim(ReduceAutotuneKey),
+    /// Key for reduce dim operations
+    ReduceDim(ReduceAutotuneKey),
     #[cfg(any(feature = "fusion", test))]
     /// Key for fused element wise operations.
     FusionElemWise(FusionElemWiseAutotuneKey),
@@ -24,8 +22,7 @@ impl Display for JitAutotuneKey {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             JitAutotuneKey::Matmul(matmul_key) => std::fmt::Display::fmt(&matmul_key, f),
-            JitAutotuneKey::SumDim(reduce_key) => std::fmt::Display::fmt(&reduce_key, f),
-            JitAutotuneKey::MeanDim(reduce_key) => std::fmt::Display::fmt(&reduce_key, f),
+            JitAutotuneKey::ReduceDim(reduce_key) => std::fmt::Display::fmt(&reduce_key, f),
             #[cfg(any(feature = "fusion", test))]
             JitAutotuneKey::FusionElemWise(reduce_key) => std::fmt::Display::fmt(&reduce_key, f),
         }
