+++ crates/burn-wgpu/src/kernel/reduce/mod.rs
@@ -1,7 +1,11 @@
+mod argmax_dim;
+mod argmin_dim;
 mod base;
+mod mean_dim;
 mod naive_reduce_shader;
 mod reduction;
 mod shared_reduce_shader;
+mod sum_dim;
 mod tune;
 
 pub use base::*;
@@ -9,3 +13,7 @@ pub use naive_reduce_shader::*;
 pub use reduction::*;
 pub use shared_reduce_shader::*;
 pub use tune::*;
+pub use sum_dim::*;
+pub use mean_dim::*;
+pub use argmin_dim::*;
+pub use argmax_dim::*;
