+++ crates/burn-wgpu/src/kernel/reduce/naive_reduce_shader.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use crate::{
     codegen::{
-        dialect::gpu::{gpu, Elem, Item, Scope, Variable, Visibility},
+        dialect::gpu::{gpu, Elem, Scope, Variable, Visibility},
         execute_dynamic, Compilation, CompilationInfo, CompilationSettings, Compiler, EagerHandle,
         InputInfo, OutputInfo, WorkgroupLaunch,
     },
@@ -12,152 +12,9 @@ use crate::{
     Runtime,
 };
 
-pub(crate) trait NaiveReduceDim: Send + Sync + 'static {
-    type Accumulator: Copy;
+use super::ReduceDimAlgorithm;
 
-    fn initialize(scope: &mut Scope, input_item: Item, output_item: Item) -> Self::Accumulator;
-
-    fn inner_loop(
-        scope: &mut Scope,
-        accumulator: Self::Accumulator,
-        current_value: Variable,
-        i: Variable,
-    );
-
-    fn assign(
-        scope: &mut Scope,
-        output: Variable,
-        accumulator: Self::Accumulator,
-        shape_reduce_dim: Variable,
-    );
-}
-
-pub(crate) struct SumDim;
-
-impl NaiveReduceDim for SumDim {
-    type Accumulator = Variable;
-
-    fn initialize(scope: &mut Scope, _input_item: Item, output_item: Item) -> Variable {
-        scope.zero(output_item)
-    }
-
-    fn inner_loop(scope: &mut Scope, accumulator: Variable, value: Variable, _i: Variable) {
-        gpu!(scope, accumulator += value);
-    }
-
-    fn assign(
-        scope: &mut Scope,
-        output: Variable,
-        accumulator: Variable,
-        _shape_reduce_dim: Variable,
-    ) {
-        let id = Variable::Id;
-        gpu!(scope, output[id] = accumulator);
-    }
-}
-
-pub(crate) struct MeanDim;
-
-impl NaiveReduceDim for MeanDim {
-    type Accumulator = Variable;
-
-    fn initialize(scope: &mut Scope, _input_item: Item, output_item: Item) -> Variable {
-        scope.zero(output_item)
-    }
-
-    fn inner_loop(scope: &mut Scope, accumulator: Variable, value: Variable, _i: Variable) {
-        gpu!(scope, accumulator += value);
-    }
-
-    fn assign(
-        scope: &mut Scope,
-        output: Variable,
-        accumulator: Variable,
-        shape_reduce_dim: Variable,
-    ) {
-        let id = Variable::Id;
-        let denominator = scope.create_local(accumulator.item());
-        gpu!(scope, denominator = cast(shape_reduce_dim));
-        gpu!(scope, accumulator = accumulator / denominator);
-        gpu!(scope, output[id] = accumulator);
-    }
-}
-
-pub(crate) struct ArgMax;
-
-impl NaiveReduceDim for ArgMax {
-    type Accumulator = (Variable, Variable);
-
-    fn initialize(scope: &mut Scope, input_item: Item, _output_item: Item) -> Self::Accumulator {
-        let max = scope.create_local(input_item);
-        let index = scope.create_local(Elem::UInt);
-        gpu!(scope, max = cast(-32767.0));
-        (max, index)
-    }
-
-    fn inner_loop(
-        scope: &mut Scope,
-        (max, index): Self::Accumulator,
-        value: Variable,
-        i: Variable,
-    ) {
-        let condition = scope.create_local(Elem::Bool);
-        gpu!(scope, condition = value > max);
-        gpu!(scope, if(condition).then(|scope| {
-            gpu!(scope, max = value);
-            gpu!(scope, index = i);
-        }));
-    }
-
-    fn assign(
-        scope: &mut Scope,
-        output: Variable,
-        (_max, index): Self::Accumulator,
-        _shape_reduce_dim: Variable,
-    ) {
-        let id = Variable::Id;
-        gpu!(scope, output[id] = index);
-    }
-}
-
-pub(crate) struct ArgMin;
-
-impl NaiveReduceDim for ArgMin {
-    type Accumulator = (Variable, Variable);
-
-    fn initialize(scope: &mut Scope, input_item: Item, _output_item: Item) -> Self::Accumulator {
-        let min = scope.create_local(input_item);
-        let index = scope.create_local(Elem::UInt);
-        gpu!(scope, min = cast(32767.0));
-        (min, index)
-    }
-
-    fn inner_loop(
-        scope: &mut Scope,
-        (min, index): Self::Accumulator,
-        value: Variable,
-        i: Variable,
-    ) {
-        let condition = scope.create_local(Elem::Bool);
-        gpu!(scope, condition = value < min);
-        gpu!(scope, if(condition).then(|scope| {
-            gpu!(scope, min = value);
-            gpu!(scope, index = i);
-        }));
-    }
-
-    fn assign(
-        scope: &mut Scope,
-        output: Variable,
-        (_min, index): Self::Accumulator,
-        _shape_reduce_dim: Variable,
-    ) {
-        let id = Variable::Id;
-        gpu!(scope, output[id] = index);
-    }
-}
-
-pub(crate) struct NaiveReduceDimComputeShader<RD: NaiveReduceDim> {
+pub(crate) struct NaiveReduceDimComputeShader<RD: ReduceDimAlgorithm> {
     tensor: Variable,
     dim: usize,
     output: Variable,
@@ -166,7 +23,7 @@ pub(crate) struct NaiveReduceDimComputeShader<RD: NaiveReduceDim> {
 
 #[derive(new)]
 pub(crate) struct NaiveReduceDimEagerKernel<
-    RD: NaiveReduceDim,
+    RD: ReduceDimAlgorithm,
     R: Runtime,
     EI: JitElement,
     EO: JitElement,
@@ -178,7 +35,7 @@ pub(crate) struct NaiveReduceDimEagerKernel<
     _elem_out: PhantomData<EO>,
 }
 
-impl<RD: NaiveReduceDim, R: Runtime, EI: JitElement, EO: JitElement> DynamicKernelSource
+impl<RD: ReduceDimAlgorithm, R: Runtime, EI: JitElement, EO: JitElement> DynamicKernelSource
     for NaiveReduceDimEagerKernel<RD, R, EI, EO>
 {
     fn source(&self) -> crate::kernel::SourceTemplate {
@@ -223,7 +80,7 @@ impl<RD: NaiveReduceDim, R: Runtime, EI: JitElement, EO: JitElement> DynamicKern
     }
 }
 
-impl<RD: NaiveReduceDim> NaiveReduceDimComputeShader<RD> {
+impl<RD: ReduceDimAlgorithm> NaiveReduceDimComputeShader<RD> {
     pub(crate) fn expand(self, scope: &mut Scope) {
         let tensor = self.tensor;
         let dim: Variable = self.dim.into();
@@ -263,7 +120,7 @@ impl<RD: NaiveReduceDim> NaiveReduceDimComputeShader<RD> {
             })
         );
 
-        let accumulator = RD::initialize(scope, tensor.item(), output.item());
+        let accumulator = RD::initialize_naive(scope, tensor.item(), output.item());
 
         gpu!(
             scope,
@@ -273,16 +130,16 @@ impl<RD: NaiveReduceDim> NaiveReduceDimComputeShader<RD> {
                 gpu!(scope, index += offset_input);
                 let value = scope.create_local(tensor.item());
                 gpu!(scope, value = tensor[index]);
-                RD::inner_loop(scope, accumulator, value, i);
+                RD::inner_loop_naive(scope, accumulator, value, i);
             })
         );
 
-        RD::assign(scope, output, accumulator, shape_input_dim);
+        RD::assign_naive(scope, output, accumulator, shape_input_dim);
     }
 }
 
 pub(crate) fn reduce_dim_naive<
-    RD: NaiveReduceDim,
+    RD: ReduceDimAlgorithm,
     R: Runtime,
     EI: JitElement,
     EO: JitElement,
@@ -314,69 +171,15 @@ pub(crate) fn reduce_dim_naive<
     output
 }
 
-/// Execute the sum dim kernel.
-pub fn sum_dim_naive<R: Runtime, E: JitElement, const D: usize>(
-    input: JitTensor<R, E, D>,
-    output: JitTensor<R, E, D>,
-    dim: usize,
-) -> JitTensor<R, E, D> {
-    reduce_dim_naive::<SumDim, R, E, E, D>(input, output, dim)
-}
-
-/// Execute the mean dim kernel.
-pub fn mean_dim_naive<R: Runtime, E: JitElement, const D: usize>(
-    input: JitTensor<R, E, D>,
-    output: JitTensor<R, E, D>,
-    dim: usize,
-) -> JitTensor<R, E, D> {
-    reduce_dim_naive::<MeanDim, R, E, E, D>(input, output, dim)
-}
-
-/// Execute the argmax kernel.
-pub fn argmax_naive<R: Runtime, E: JitElement, I: JitElement, const D: usize>(
-    input: JitTensor<R, E, D>,
-    dim: usize,
-) -> JitTensor<R, I, D> {
-    let mut shape_out = input.shape.clone();
-    shape_out.dims[dim] = 1;
-    let num_elems = shape_out.num_elements();
-    let buffer = input.client.empty(num_elems * core::mem::size_of::<I>());
-    let output = JitTensor::new(
-        input.client.clone(),
-        input.device.clone(),
-        shape_out,
-        buffer,
-    );
-    reduce_dim_naive::<ArgMax, R, E, I, D>(input, output, dim)
-}
-
-/// Execute the argmin kernel.
-pub fn argmin_naive<R: Runtime, E: JitElement, I: JitElement, const D: usize>(
-    input: JitTensor<R, E, D>,
-    dim: usize,
-) -> JitTensor<R, I, D> {
-    let mut shape_out = input.shape.clone();
-    shape_out.dims[dim] = 1;
-    let num_elems = shape_out.num_elements();
-    let buffer = input.client.empty(num_elems * core::mem::size_of::<I>());
-    let output = JitTensor::new(
-        input.client.clone(),
-        input.device.clone(),
-        shape_out,
-        buffer,
-    );
-    reduce_dim_naive::<ArgMin, R, E, I, D>(input, output, dim)
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{
-        kernel::reduce::{argmax_naive, init_reduce_output, mean_dim_naive},
-        tests::{ReferenceBackend, TestBackend},
+        kernel::reduce::{init_reduce_output, ArgMax, ArgMin, MeanDim, SumDim},
+        tests::{ReferenceBackend, TestBackend, TestRuntime},
     };
     use burn_tensor::{ops::IntTensorOps, Data, Distribution, Int, Shape, Tensor};
 
-    use super::sum_dim_naive;
+    use super::reduce_dim_naive;
 
     #[test]
     fn reduction_sum_dim_should_work_with_multiple_invocations() {
@@ -387,26 +190,67 @@ mod tests {
         let reduce_dim = 1;
         let output = init_reduce_output(&tensor.clone().into_primitive(), reduce_dim);
 
-        let val = Tensor::<TestBackend, 2>::from_primitive(sum_dim_naive(
-            tensor.into_primitive(),
-            output,
-            reduce_dim,
-        ));
+        let val =
+            Tensor::<TestBackend, 2>::from_primitive(reduce_dim_naive::<
+                SumDim,
+                TestRuntime,
+                f32,
+                f32,
+                2,
+            >(
+                tensor.into_primitive(), output, reduce_dim
+            ));
         let val_ref = tensor_ref.sum_dim(1);
 
         val_ref.into_data().assert_approx_eq(&val.into_data(), 3);
     }
 
     #[test]
-    fn reduction_args_dim_should_work_with_multiple_invocations() {
+    fn reduction_argmin_dim_should_work_with_multiple_invocations() {
         let tensor =
             Tensor::<TestBackend, 2>::random([6, 1024], Distribution::Default, &Default::default());
         let tensor_ref =
             Tensor::<ReferenceBackend, 2>::from_data(tensor.to_data(), &Default::default());
+        let reduce_dim = 1;
+        let output = init_reduce_output(&tensor.clone().into_primitive(), reduce_dim);
 
-        let val =
-            Tensor::<TestBackend, 2, Int>::from_primitive(argmax_naive(tensor.into_primitive(), 1));
-        let val_ref = tensor_ref.argmax(1);
+        let val = Tensor::<TestBackend, 2, Int>::from_primitive(reduce_dim_naive::<
+            ArgMin,
+            TestRuntime,
+            f32,
+            i32,
+            2,
+        >(
+            tensor.into_primitive(),
+            output,
+            reduce_dim,
+        ));
+        let val_ref = tensor_ref.argmin(reduce_dim);
+
+        assert_eq!(val_ref.into_data().convert(), val.into_data());
+    }
+
+    #[test]
+    fn reduction_argmax_dim_should_work_with_multiple_invocations() {
+        let tensor =
+            Tensor::<TestBackend, 2>::random([6, 1024], Distribution::Default, &Default::default());
+        let tensor_ref =
+            Tensor::<ReferenceBackend, 2>::from_data(tensor.to_data(), &Default::default());
+        let reduce_dim = 1;
+        let output = init_reduce_output(&tensor.clone().into_primitive(), reduce_dim);
+
+        let val = Tensor::<TestBackend, 2, Int>::from_primitive(reduce_dim_naive::<
+            ArgMax,
+            TestRuntime,
+            f32,
+            i32,
+            2,
+        >(
+            tensor.into_primitive(),
+            output,
+            reduce_dim,
+        ));
+        let val_ref = tensor_ref.argmax(reduce_dim);
 
         assert_eq!(val_ref.into_data().convert(), val.into_data());
     }
@@ -419,8 +263,13 @@ mod tests {
 
         let summed_tensor = TestBackend::int_empty(summed_shape, &Default::default());
 
-        let val =
-            Tensor::<TestBackend, 1, Int>::from_primitive(sum_dim_naive(tensor, summed_tensor, 0));
+        let val = Tensor::<TestBackend, 1, Int>::from_primitive(reduce_dim_naive::<
+            SumDim,
+            TestRuntime,
+            i32,
+            i32,
+            1,
+        >(tensor, summed_tensor, 0));
 
         let sum_as_data = Data::from([10]);
         val.into_data().assert_approx_eq(&sum_as_data, 1);
@@ -434,8 +283,13 @@ mod tests {
 
         let mean_tensor = TestBackend::int_empty(mean_shape, &Default::default());
 
-        let val =
-            Tensor::<TestBackend, 1, Int>::from_primitive(mean_dim_naive(tensor, mean_tensor, 0));
+        let val = Tensor::<TestBackend, 1, Int>::from_primitive(reduce_dim_naive::<
+            MeanDim,
+            TestRuntime,
+            i32,
+            i32,
+            1,
+        >(tensor, mean_tensor, 0));
 
         // Mean calculation truncates to an integer
         let mean_as_data = Data::from([2]);
