+++ crates/burn-wgpu/src/kernel/reduce/tune/mod.rs
@@ -1,8 +1,8 @@
 mod base;
 mod key;
-mod mean_dim;
-mod sum_dim;
+// mod mean_dim;
+// mod sum_dim;
 
 pub use key::*;
-pub use mean_dim::*;
-pub use sum_dim::*;
+// pub use mean_dim::*;
+// pub use sum_dim::*;
